Fix non-POSIX behavior
Since dash (and POSIX shell as well) doesn't support multiline variable
definitions as bash/ksh so FILELIST variable is set as single string.
Also variables assignment doesn't work as they was expected with commands
executed with 'eval' in dash/ksh (but works with bash 5.2).
For clarity the 'mkdeploy' invocation is rewritten with other ugly
but working 'eval' command.